abstracted confetti to a shared module, linked positivity setting
