+++ src/App.js
@@ -134,7 +134,8 @@ const App = (props) => {
     today,
     data,
     streak,
-    missed
+    missed,
+    positivity
   };
 
   const settingsProps = {
