+++ src/shared/confetti.js
@@ -1,0 +1,24 @@
+import confetti from 'canvas-confetti';
+
+export const fireworks = () => {
+  const duration = 3 * 1000;
+  const animationEnd = Date.now() + duration;
+  const defaults = { startVelocity: 30, spread: 360, ticks: 60, zIndex: 100 };
+  
+  const randomInRange = (min, max) => {
+    return Math.random() * (max - min) + min;
+  };
+  
+  const interval = setInterval(() => {
+    const timeLeft = animationEnd - Date.now();
+  
+    if (timeLeft <= 0) {
+      return clearInterval(interval);
+    }
+  
+    const particleCount = 250 * (timeLeft / duration);
+    // since particles fall down, start a bit higher than random
+    confetti(Object.assign({}, defaults, { particleCount, origin: { x: randomInRange(0.1, 0.3), y: Math.random() - 0.2 } }));
+    confetti(Object.assign({}, defaults, { particleCount, origin: { x: randomInRange(0.7, 0.9), y: Math.random() - 0.2 } }));
+  }, 250);
+};
